add array of product identifiers
you can get an ordered list of product descriptions and an unordered
dictionary with prices which is pretty much useless on its own, unless
you want to do management yourself which renders a lot of different code
useless. so just add a list of ordered product identifiers (which
matches the indices in the list of descriptions) which can be used to
initiate a purchase, still allowing for manual management of the list
but adding a better way to manage it automatically.