don't include review code if REVIEW_ALLOWED is not defined
previously we would always include the code, even if review was not
allowed (e.g. has a value of 0).
making sure that this code does not make it into a released binary is a
good thing for security as we can't rely on the compiler to do this for
us.
unfortunately in theory REVIEW_ALLOWED could be a more complex statement
so if it is defined at all we include the code and evalute the
expression.